use node v.14 for client
Tagged the client docker image into node-v14 to
solve an issue related to ssl library changes. With
newer node versions it is required to use --openssl-legacy-provider
to make the project setup work.

The issue is related to Webpack 4
(https://github.com/webpack/webpack/issues/14532) and would
require updating the tooling to use a newer webpack version.